requirements.txt
torch==2.10.0+cu130
numpy==2.2.6
scipy==1.17.0
matplotlib==3.10.8
PyYAML==6.0.3
